Rename integrations so they're easier to understand
