+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Remoting/Server/ProcessMessageIntegration.cs
@@ -1,4 +1,4 @@
-// <copyright file="SoapServerFormatterSinkIntegration.cs" company="Datadog">
+// <copyright file="ProcessMessageIntegration.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -13,10 +13,10 @@ using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Logging;
 using Datadog.Trace.Propagators;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Remoting
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Remoting.Server
 {
     /// <summary>
-    /// System.Runtime.Remoting.Channels.SoapClientFormatterSink.ProcessMessage calltarget instrumentation
+    /// System.Runtime.Remoting.Channels.IServerChannelSink.ProcessMessage calltarget instrumentation
     /// </summary>
     [InstrumentMethod(
         AssemblyName = "System.Runtime.Remoting",
@@ -39,9 +39,9 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Remoting
     [Browsable(false)]
     [EditorBrowsable(EditorBrowsableState.Never)]
     // ReSharper disable once InconsistentNaming
-    public class SoapServerFormatterSinkIntegration
+    public class ProcessMessageIntegration
     {
-        private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(SoapServerFormatterSinkIntegration));
+        private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(ProcessMessageIntegration));
 
         /// <summary>
         /// OnMethodBegin callback
